+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aave/react
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @aave/graphql@0.3.1
+  - @aave/client@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes